+++ ticketkatjuscha/src/main/java/the/dancing/company/ticketkatjuscha/TicketGenerator.java
@@ -8,7 +8,7 @@ import org.javatuples.Pair;
 import com.itextpdf.text.DocumentException;
 
 public interface TicketGenerator {
-    public void generate(String code, String checkCode, String ticketOwnerName, Pair<String, String> seat, int price, OutputStream output) throws IOException, DocumentException;
+    public void generate(String code, String checkCode, String ticketOwnerName, Pair<String, String> seat, double price, OutputStream output) throws IOException, DocumentException;
 
     public String getFileNameExtension();
 }
